requirements.txt
Flask
pymysql
Faker
python-dateutil
